[807] Move the REG_CONSTANT to constant.py.
