+++ examples/rolling_process_data/workflow.py
@@ -6,7 +6,7 @@ import fire
 import pickle
 
 from datetime import datetime
-from qlib.config import REG_CN
+from qlib.constant import REG_CN
 from qlib.data.dataset.handler import DataHandlerLP
 from qlib.utils import init_instance_by_config
 from qlib.tests.data import GetData
