+++ qlib/constant.py
@@ -1,0 +1,6 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+# REGION CONST
+REG_CN = "cn"
+REG_US = "us"
