+++ scripts/README.md
@@ -62,7 +62,8 @@ qlib.init(provider_uri=provider_uri, region=REG_US)
 
 ```python
 import qlib
-from qlib.config import REG_CN
+from qlib.constant import REG_CN
+
 provider_uri = "~/.qlib/qlib_data/cn_data"  # target_dir
 qlib.init(provider_uri=provider_uri, region=REG_CN)
 ```
